requirements.txt
accelerate==1.2.1
transformers==4.47.1
diffusers==0.31.0
ftfy==6.3.1
# albumentations==1.3.0
opencv-python==4.10.0.84
einops==0.8.0
pytorch-lightning==2.5.0.post0
bitsandbytes==0.45.0
prodigyopt==1.1.1
lion-pytorch==0.2.3
schedulefree==1.4
tensorboard==2.18.0
safetensors==0.4.5
hpsv2==1.2.0
# gradio==3.16.2
# Breaking changes with newer versions
altair==4.2.2 
easygui==0.98.3
toml==0.10.2
voluptuous==0.15.2
# Breaking changes with newer versions
huggingface-hub==0.25.2
matplotlib==3.10.0
scipy==1.14.1
# for Image utils
imagesize==1.4.1
# for BLIP captioning
# requests==2.28.2
# timm==0.6.12
# fairscale==0.4.13
# for WD14 captioning (tensorflow)
# tensorflow==2.10.1
# for WD14 captioning (onnx)
# onnx==1.15.0
# onnxruntime-gpu==1.17.1
# onnxruntime==1.17.1
# for cuda 12.1(default 11.8)
# onnxruntime-gpu --extra-index-url https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/onnxruntime-cuda-12/pypi/simple/

# this is for onnx: 
# protobuf==3.20.3
# open clip for SDXL
# open-clip-torch==2.20.0
# For logging
rich==13.9.4
# for T5XXL tokenizer (SD3/FLUX)
sentencepiece==0.2.0
# for kohya_ss library
-e .
